new sobel version avx512
test version, maybe not working

+++ include/common.h
@@ -19,7 +19,7 @@ f64 mean(f64 *a, u64 n);
 f64 stddev(f64 *a, u64 n);
 
 
-void grayscale_weighted(u8 *bigFrame, u8 *smallFrame);
-void grayscale_sampled(u8 *bigFrame, u8 *smallFrame);
+void grayscale_weighted(u8 *bigFrame, i16 *smallFrame);
+void grayscale_sampled(u8 *bigFrame, i16 *smallFrame);
 
-void Biggerize(u8 *smallFrame, u8 *bigFrame);
+void Biggerize(i16 *smallFrame, u8 *bigFrame);
